Time: 0 ms (100%), Space: 42.2 MB (26.56%) - LeetHub

+++ 0238-product-of-array-except-self/0238-product-of-array-except-self.cpp
@@ -2,8 +2,7 @@ class Solution {
 public:
     vector<int> productExceptSelf(vector<int>& nums) {
         vector<int> pre(nums.size()),suf(nums.size());
-        long long cp=1;
-
+        int cp=1;
         for(int i=0;i<nums.size();i++){
             pre[i]=cp;
             cp*=nums[i];
